api/requirements.txt
django==1.8
djangorestframework
django-rest-auth[with_social]
django-rest-swagger
psycopg2
django-cuser==2014.9.28
sqlparse
mock
django-filter
django-cors-headers==1.1.0
